requirements.txt
absl-py==2.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
fastapi==0.135.2
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
prettytable==3.17.0
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==1.0.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
waitress==3.0.2
wcwidth==0.6.0
Werkzeug==3.1.7
zipp==3.23.0
